+++ src/main/java/com/example/gimmegonghakauth/completed/service/exception/FileException.java
@@ -1,4 +1,4 @@
-package com.example.gimmegonghakauth.exception;
+package com.example.gimmegonghakauth.completed.service.exception;
 
 public class FileException extends RuntimeException {
