+++ src/test/java/com/example/gimmegonghakauth/user/service/UserServiceTest.java
@@ -3,9 +3,8 @@ package com.example.gimmegonghakauth.user.service;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.example.gimmegonghakauth.exception.UserNotFoundException;
+import com.example.gimmegonghakauth.user.service.exception.UserNotFoundException;
 import com.example.gimmegonghakauth.user.domain.UserDomain;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
